Rename out3SimpPlot to out1Peak and update documentation
