+++ avaframe/tests/test_com1DFA.py
@@ -10,7 +10,7 @@ import numpy as np
 import os
 from avaframe.com1DFA import com1DFA
 from avaframe.in3Utils import cfgUtils
-from avaframe.out3SimpPlot import outPlotAllPeak as oP
+from avaframe.out1Peak import outPlotAllPeak as oP
 from avaframe.log2Report import generateReport as gR
 from avaframe.in3Utils import initializeProject as initProj
 from benchmarks import simParameters
